requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
cohere==5.14.2
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
packaging==24.2
pydantic==2.11.2
pydantic_core==2.33.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
starlette==0.46.1
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
wheel==0.45.1
uvicorn